Prevent pointer derefence crash on failure to load libva library
vaDisplayIsValidPFN is a special case in VASharingFunctions. Even
if libva was not loaded this function could be called thru
VASharingFunctions::isValidVaDisplay which is the case in NEO
driver: see VaSharingContextBuilder::finalizeProperties. Mind that
libva could be of different versions: 1.x, 2.x, etc. Which means
that situations possible that application passed valid VADisplay
for libva-2.x, but NEO driver tries to load libva-1.x.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>