Add support of GPU preemption (v0.6).
Use printDebugString() & DebugManager.log() to print debug message
at the same time. It makes the code looks redundant and silly a little bit,
but it has a reason. DebugManager.log() only works in the debug version
whose size is about 500MB, some users complain this and it’s impossible
for some resource-limited devices, but they need the log. So in addition
to the DebugManager.log(), I use  printDebugString() too with configuration
file igdrcl.config.