Fix calculation in encodeGreaterThanPredicate()
encodeGreaterThanPredicate() computes *firstOperand > secondOperand and
stores the result in MI_PREDICATE_RESULT where firstOperand is an
device memory address.

To calculate the "greater than" operation in the device,
(secondOperand - *firstOperand) needs to be used, and if the carry flag
register is set, then (*firstOperand) is greater than secondOperand.

The order of operands in the subtraction (secondOperand - *firstOperand)
in encodeGreaterThanPredicate() is incorrect. Transpose the
operands.

Change-Id: Ie7fee32af338a1a3948abc78cf8332e0be8595ac
Signed-off-by: Sebastian Sanchez <sebastian.sanchez@intel.com>