Some debian package manager tweaks
By default, Ubuntu or Debian based "apt" or "apt-get" system installs
recommended but not suggested packages.

By passing "--no-install-recommends" option, the user lets apt-get know
not to consider recommended packages as a dependency to install.

This results in smaller downloads and installation of packages.

Original pull request #268 was modified to fix dependencies on
Ubuntu 18.04 and newer.

Signed-off-by: Pratik Raj <rajpratik71@gmail.com>
Signed-off-by: Jacek Danecki <jacek.danecki@intel.com>

https://github.com/intel/compute-runtime/pull/268

Change-Id: I5a5de36102c5d3d3a42def30404078c485705b8a