Fixed Image size used in Copy From/to memory for null regions
- When peforming Image Copy from/to memory, the size of the image and thus the
memory should be expected to be at maximum the size of the
image width, height, & depth in the user's descriptor.
- When a user passes in a nullptr for the destination or source region the
width, height, & depth of the image region should match the image descriptor
that was passed in by the user at image create.
- The Physical image allocated may be larger than the user requested.
This new size cannot be used in the Copy to/From memory otherwise device data
corruption will occur.

Change-Id: I950f6fd1d5ddc9911dd235f84e1cf00618e4e2a3
Signed-off-by: Spruit, Neil R <neil.r.spruit@intel.com>