Create L0 list and module with correct NEO device on explicit mode
When implicit scaling is not used, a root device in L0 defaults to
sub-device 0. In that case, all allocations need to be made against
that sub-device.

For list and module creation, we were passing the root NEO device
instead, and we were allocating their internal allocations using
the least-occupied bank, which meant we were allocating alternately
between the two sub-devices.

This patch makes sure the NEO device used on explicit mode for
the root device is sub-device 0.

Signed-off-by: Jaime Arteaga <jaime.a.arteaga.molina@intel.com>