Disable SET_PAIR by default.
To use, applications need to set EnableSetPair=1 explicitly.

When disabled, implicit scaling allocations require two IPC handles
to be exchanged with other processes using the zexMemGetIpcHandles
APIs.

When enabled, implicit scaling allocations only require one IPC
handle to be exchanged with other process using the zeMemGetIpcHandle
APIs. This is only available when allocation is imported in a different
device than the one in the exporter.

Signed-off-by: Jaime Arteaga <jaime.a.arteaga.molina@intel.com>