linux/clgl: Add support for MESA GL interop functions
This is just the header and function prototypes and wrapper so far.

The interop header comes from MESA repository at
https://gitlab.freedesktop.org/mesa/mesa
and can be found under mesa/include/GL/mesa_glinterop.h

It has been stripped of the WGL prototypes since we don't need
those here and they cause a type conflict in this project.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>