fix: Track Sub Device Hierarchy for use by Metrics Library
Related-To: LOCI-4819

- When communicating with the Metrics Libraries, if one is using a
specific sub device, then the correct physical sub device index and
number of sub devices must be used.

- When Affinity Mask is set, this hierarchy information is lost,
therefore the sub device hierarchy data is stored in a map to each
"device" to be returned to the user allowing for the correct sub device
information to be passed to the metrics library.

Signed-off-by: Spruit, Neil R <neil.r.spruit@intel.com>
Source: 5c75449508cd8d73916baebce0abb9a73ed8daca