feature: report multi-hop fabric connections
In additon to physical connections we should also report multi-hop
logical connections (MDFI + XeLink) as have positive bandwidth.

Use a modified BFS algorithm to try to find a path between fabric
vertices that are not directly connected together because the KMD always
try to use MDFI link first, then go to XeLink.

Multi-hop connections are bi-directional but might not be symmetric, so
for every pair of vertices A & B that are not directly connected, we
need to try to find both `A -> B` and `B -> A`.

Related-To: GSD-7126

Signed-off-by: Wenbin Lu <wenbin.lu@intel.com>
Source: a0faad6558ef41cc64fbccc76d01fe4ef6a541be