fix: set kernel LWS prior to obtaining indirect heaps
When calculating the size of the indirect object heap,
the local work group size from kernel implicit args is taken into account.
If the LWS is not set before this calculation,
it can lead to insufficient ioh allocation size.
Such a problem is seen when local ids are generated by the runtime
and then written to ioh. The write fails due to lack of space in the allocation.

Related-To: IGC-7708
Signed-off-by: Fabian Zwolinski <fabian.zwolinski@intel.com>