fix: ocloc fatbinary handle -out_dir without -output
Currently if we pass `-out_dir` in ocloc fatbinary cmdline, e.g.:
./ocloc compile -file vector.cl ... -out_dir ../exampler_dir
ocloc sets the name for the output file from the `-output` parameter.
But as the `-output` parameter is not provided,
the file name will be empty and won't be written to disk.

This patch adds support for a scenario
where you pass `-out_dir` without `-output`.
In this case, the file will have default name - input file's base name.

Related-To: NEO-10603
Signed-off-by: Fabian Zwolinski <fabian.zwolinski@intel.com>