feature: finalizes 2d block load/store query
Finalized the implementation of the 2d block
load/store extension query. Namely, this adds
the extension info to the `DriverHandleImp`.

Also fixed support matrix in the current
implementation that incorrectly includes
MTL and ARL in the supported products. ULTs
and the implementation have both been changed
to match the true support matrix.

Related-To: NEO-11592
Signed-off-by: Jack Myers <jack.myers@intel.com>